lotteryRound function moved to LotteryService
